add shared jest esm packages list
